[ABI] Added AArch64 JIT pipelines and new JIT abstractions
  * Version bump to 0.11.0
  * Pipeline compiler now uses full SIMD abstractions
  * Pipeline compiler now uses PixelPredicate for all fetches
  * Composition for AArch64 has a platform specific code-path
  * Improved performance of radial gradients by 30-45% on X86
  * Improved performance of bilinear texture fetches by 10%
  * Radial gradients now support both center and focal point radius
  * Conic gradients now support repeat parameter and honor PAD and
    REPEAT modes (compatibility with CSS repeating-conic-gradient)
    (this feature is experimental and needs further enhancements)
  * Added new tests that can compare portable and JIT pipelines
  * Added a hardened build to CI (uses C++ hardening options)
  * Pipeline tests now run on CI for all supported backends and
    their sub-targets (SSE2, SSE3, SSSE4, SSE4.X, AVX, AVX2, AVX-512,
    ASIMD)
  * Pipeline tests can now set alpha to opaque or semi-transparent
  * Pipeline tests can now use explicit style API and state-full
    style API such as setFillStyle() and setStrokeStyle()
  * Pipeline tests can now test stroking functionality in addition
    to filling
  * Fixed many pipeline bugs thanks to the new testing architecture
  * Fixed rendering of A8 patterns when rendered into XRGB32/PRGB32
    buffers
  * Fixed MSAN issues (MSAN now enabled on CI as well)
  * Fixed TSAN issues (TSAN now enabled on CI as well)
  * Fixed a bug that caused BLPattern not being freed with using
    setFillStyle() and setStrokeStype() with BLPattern on a
    multi-threaded rendering context (BLPattern being confused with
    BLImage at the free phase)

+++ src/blend2d/opentype/otname.cpp
@@ -68,7 +68,7 @@ static BLResult convertNameStringToUtf8(BLString& dst, BLArrayView<uint8_t> src,
       uint32_t uc;
       BL_PROPAGATE(srcReader.next<Unicode::IOFlags::kUnaligned | Unicode::IOFlags::kByteOrderBE | Unicode::IOFlags::kStrict>(uc));
 
-      nullTerminatorCount += (uc == 0);
+      nullTerminatorCount += size_t(uc == 0);
       BL_PROPAGATE(dstWriter.writeUnsafe(uc));
     }
   }
